jw - moved class Entry and Folder out of FullTree class
