+++ app/src/main/java/com/github/mobile/ui/code/RepositoryCodeFragment.java
@@ -38,8 +38,8 @@ import android.widget.TextView;
 import com.github.kevinsawicki.wishlist.ViewUtils;
 import com.github.mobile.R;
 import com.github.mobile.core.code.FullTree;
-import com.github.mobile.core.code.FullTree.Entry;
-import com.github.mobile.core.code.FullTree.Folder;
+import com.github.mobile.core.code.TreeEntry.Entry;
+import com.github.mobile.core.code.TreeFolder.Folder;
 import com.github.mobile.core.code.RefreshTreeTask;
 import com.github.mobile.core.ref.RefUtils;
 import com.github.mobile.ui.DialogFragment;
